refactor(router): split resolvers & guards into separate files
